pyproject.toml
[build-system]
requires = ["hatchling>=1.0.0,<2"]
build-backend = "hatchling.build"

[project]
name = "pdf-format-analyzer"
dynamic = ["version"]
description = "Detect and fix LaTeX formatting issues in PDFs using vision LLM analysis via GitHub Copilot SDK"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Arthur Freitas Ramos" }]
keywords = ["pdf", "latex", "formatting", "copilot", "vision", "linting"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: LaTeX",
]
dependencies = [
    "github-copilot-sdk>=0.1.23,<0.2",
    "typer>=0.9.0,<1",
    "rich>=13.0.0,<14",
    "pydantic>=2.0.0,<3",
    "PyMuPDF>=1.24.0",
    "Pillow>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/ArthurFreitasRamos/pdf-format-analyzer"
Repository = "https://github.com/ArthurFreitasRamos/pdf-format-analyzer"
Issues = "https://github.com/ArthurFreitasRamos/pdf-format-analyzer/issues"

[project.optional-dependencies]
dev = ["pytest>=8.0,<10", "pytest-asyncio>=0.23,<2", "pytest-cov>=5.0,<7", "ruff<1"]

[project.scripts]
pfa = "pdf_format_analyzer.cli:app"

[tool.hatch.version]
path = "src/pdf_format_analyzer/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_format_analyzer"]

[tool.hatch.build.targets.sdist]
include = ["src/pdf_format_analyzer"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
